Struct google_dfareporting2d1::AccountActiveAdSummaryMethods

pub struct AccountActiveAdSummaryMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on accountActiveAdSummary resources. It is not used directly, but through the Dfareporting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_dfareporting2d1 as dfareporting2d1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2d1::Dfareporting;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`
// to build up your call.
let rb = hub.account_active_ad_summaries();

Methods

impl<'a, C, A> AccountActiveAdSummaryMethods<'a, C, A>

fn get(&self, profile_id: &str, summary_account_id: &str) -> AccountActiveAdSummaryGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the account's active ad summary by account ID.

Arguments

profileId - User profile ID associated with this request.

summaryAccountId - Account ID.