Struct google_dfareporting2d1::CompatibleFields

pub struct CompatibleFields {
    pub kind: Option<String>,
    pub cross_dimension_reach_report_compatible_fields: Option<CrossDimensionReachReportCompatibleFields>,
    pub reach_report_compatible_fields: Option<ReachReportCompatibleFields>,
    pub path_to_conversion_report_compatible_fields: Option<PathToConversionReportCompatibleFields>,
    pub floodlight_report_compatible_fields: Option<FloodlightReportCompatibleFields>,
    pub report_compatible_fields: Option<ReportCompatibleFields>,
}

Represents a response to the queryCompatibleFields method.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

compatible fields query reports (response)

Fields

kind

The kind of resource this is, in this case dfareporting#compatibleFields.

cross_dimension_reach_report_compatible_fields

Contains items that are compatible to be selected for a report of type "CROSS_DIMENSION_REACH".

reach_report_compatible_fields

Contains items that are compatible to be selected for a report of type "REACH".

path_to_conversion_report_compatible_fields

Contains items that are compatible to be selected for a report of type "PATH_TO_CONVERSION".

floodlight_report_compatible_fields

Contains items that are compatible to be selected for a report of type "FLOODLIGHT".

report_compatible_fields

Contains items that are compatible to be selected for a report of type "STANDARD".