Struct google_dfareporting2d1::Creative

pub struct Creative {
    pub artwork_type: Option<String>,
    pub latest_trafficked_creative_id: Option<String>,
    pub counter_custom_events: Option<Vec<CreativeCustomEvent>>,
    pub advertiser_id: Option<String>,
    pub active: Option<bool>,
    pub timer_custom_events: Option<Vec<CreativeCustomEvent>>,
    pub rendering_id_dimension_value: Option<DimensionValue>,
    pub fs_command: Option<FsCommand>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub studio_trafficked_creative_id: Option<String>,
    pub archived: Option<bool>,
    pub override_css: Option<String>,
    pub allow_script_access: Option<bool>,
    pub required_flash_version: Option<i32>,
    pub exit_custom_events: Option<Vec<CreativeCustomEvent>>,
    pub rendering_id: Option<String>,
    pub video_description: Option<String>,
    pub third_party_backup_image_impressions_url: Option<String>,
    pub version: Option<i32>,
    pub creative_field_assignments: Option<Vec<CreativeFieldAssignment>>,
    pub background_color: Option<String>,
    pub custom_key_values: Option<Vec<String>>,
    pub id_dimension_value: Option<DimensionValue>,
    pub auto_advance_images: Option<bool>,
    pub size: Option<Size>,
    pub type_: Option<String>,
    pub total_file_size: Option<String>,
    pub third_party_rich_media_impressions_url: Option<String>,
    pub studio_advertiser_id: Option<String>,
    pub creative_assets: Option<Vec<CreativeAsset>>,
    pub backup_image_features: Option<Vec<String>>,
    pub html_code_locked: Option<bool>,
    pub redirect_url: Option<String>,
    pub click_tags: Option<Vec<ClickTag>>,
    pub backup_image_target_window: Option<TargetWindow>,
    pub required_flash_plugin_version: Option<String>,
    pub backup_image_reporting_label: Option<String>,
    pub compatibility: Option<Vec<String>>,
    pub ad_parameters: Option<String>,
    pub ad_tag_keys: Option<Vec<String>>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub convert_flash_to_html5: Option<bool>,
    pub video_duration: Option<f32>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub backup_image_click_through_url: Option<String>,
    pub studio_creative_id: Option<String>,
    pub authoring_tool: Option<String>,
    pub third_party_urls: Option<Vec<ThirdPartyTrackingUrl>>,
    pub companion_creatives: Option<Vec<String>>,
    pub html_code: Option<String>,
    pub ssl_compliant: Option<bool>,
    pub commercial_id: Option<String>,
    pub skippable: Option<bool>,
}

Contains properties of a Creative.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert creatives (request|response)

get creatives (response)

list creatives (none)

update creatives (request|response)

patch creatives (request|response)

Fields

artwork_type

Type of artwork used for the creative. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

latest_trafficked_creative_id

Latest Studio trafficked creative ID associated with rich media and VPAID creatives. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

counter_custom_events

List of counter events configured for the creative. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

advertiser_id

Advertiser ID of this creative. This is a required field. Applicable to all creative types.

active

Whether the creative is active. Applicable to all creative types.

timer_custom_events

List of timer events configured for the creative. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

rendering_id_dimension_value

Dimension value for the rendering ID of this creative. This is a read-only field. Applicable to all creative types.

fs_command

OpenWindow FSCommand of this creative. This lets the SWF file communicate with either Flash Player or the program hosting Flash Player, such as a web browser. This is only triggered if allowScriptAccess field is true. Applicable to the following creative types: FLASH_INPAGE.

id

ID of this creative. This is a read-only, auto-generated field. Applicable to all creative types.

account_id

Account ID of this creative. This field, if left unset, will be auto-generated for both insert and update operations. Applicable to all creative types.

studio_trafficked_creative_id

Studio trafficked creative ID associated with rich media and VPAID creatives. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

archived

Whether the creative is archived. Applicable to all creative types.

override_css

Override CSS value for rich media creatives. Applicable to the following creative types: all RICH_MEDIA.

allow_script_access

Whether script access is allowed for this creative. This is a read-only and deprecated field which will automatically be set to true on update. Applicable to the following creative types: FLASH_INPAGE.

required_flash_version

The internal Flash version for this creative as calculated by DoubleClick Studio. This is a read-only field. Applicable to the following creative types: FLASH_INPAGE, ENHANCED_BANNER, all RICH_MEDIA, and all VPAID.

exit_custom_events

List of exit events configured for the creative. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

rendering_id

ID of current rendering version. This is a read-only field. Applicable to all creative types.

video_description

Description of the video ad. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

third_party_backup_image_impressions_url

Third-party URL used to record backup image impressions. Applicable to the following creative types: all RICH_MEDIA

version

The version number helps you keep track of multiple versions of your creative in your reports. The version number will always be auto-generated during insert operations to start at 1. For tracking creatives the version cannot be incremented and will always remain at 1. For all other creative types the version can be incremented only by 1 during update operations. In addition, the version will be automatically incremented by 1 when undergoing Rich Media creative merging. Applicable to all creative types.

creative_field_assignments

Creative field assignments for this creative. Applicable to all creative types.

background_color

The 6-character HTML color code, beginning with #, for the background of the window area where the Flash file is displayed. Default is white. Applicable to the following creative types: FLASH_INPAGE.

custom_key_values

Custom key-values for a Rich Media creative. Key-values let you customize the creative settings of a Rich Media ad running on your site without having to contact the advertiser. You can use key-values to dynamically change the look or functionality of a creative. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

id_dimension_value

Dimension value for the ID of this creative. This is a read-only field. Applicable to all creative types.

auto_advance_images

Whether images are automatically advanced for enhanced image creatives. Applicable to the following creative types: ENHANCED_IMAGE.

size

Size associated with this creative. When inserting or updating a creative either the size ID field or size width and height fields can be used. This is a required field when applicable; however for IMAGE and FLASH_INPAGE creatives, if left blank, this field will be automatically set using the actual size of the associated image assets. Applicable to the following creative types: ENHANCED_BANNER, ENHANCED_IMAGE, FLASH_INPAGE, HTML5_BANNER, IMAGE, and all RICH_MEDIA.

type_

Type of this creative.This is a required field. Applicable to all creative types.

total_file_size

Combined size of all creative assets. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

third_party_rich_media_impressions_url

Third-party URL used to record rich media impressions. Applicable to the following creative types: all RICH_MEDIA

studio_advertiser_id

Studio advertiser ID associated with rich media and VPAID creatives. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

creative_assets

Assets associated with a creative. Applicable to all but the following creative types: INTERNAL_REDIRECT, INTERSTITIAL_INTERNAL_REDIRECT, and REDIRECT

backup_image_features

List of feature dependencies that will cause a backup image to be served if the browser that serves the ad does not support them. Feature dependencies are features that a browser must be able to support in order to render your HTML5 creative asset correctly. This field is initially auto-generated to contain all features detected by DCM for all the assets of this creative and can then be modified by the client. To reset this field, copy over all the creativeAssets' detected features. Applicable to the following creative types: ENHANCED_BANNER and HTML5_BANNER.

html_code_locked

Whether HTML code is DCM-generated or manually entered. Set to true to ignore changes to htmlCode. Applicable to the following creative types: FLASH_INPAGE and HTML5_BANNER.

redirect_url

URL of hosted image or another ad tag. This is a required field when applicable. Applicable to the following creative types: INTERNAL_REDIRECT, INTERSTITIAL_INTERNAL_REDIRECT, and REDIRECT

click_tags

Click tags of the creative. For ENHANCED_BANNER, FLASH_INPAGE, and HTML5_BANNER creatives, this is a subset of detected click tags for the assets associated with this creative. After creating a flash asset, detected click tags will be returned in the creativeAssetMetadata. When inserting the creative, populate the creative clickTags field using the creativeAssetMetadata.clickTags field. For ENHANCED_IMAGE creatives, there should be exactly one entry in this list for each image creative asset. A click tag is matched with a corresponding creative asset by matching the clickTag.name field with the creativeAsset.assetIdentifier.name field. Applicable to the following creative types: ENHANCED_BANNER, ENHANCED_IMAGE, FLASH_INPAGE, HTML5_BANNER.

backup_image_target_window

Target window for backup image. Applicable to the following creative types: ENHANCED_BANNER, FLASH_INPAGE, and HTML5_BANNER.

required_flash_plugin_version

The minimum required Flash plugin version for this creative. For example, 11.2.202.235. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

backup_image_reporting_label

Reporting label used for HTML5 banner backup image. Applicable to the following creative types: ENHANCED_BANNER.

compatibility

Compatibilities associated with this creative. This is a read-only field. WEB and WEB_INTERSTITIAL refer to rendering either on desktop or on mobile devices for regular or interstitial ads, respectively. APP and APP_INTERSTITIAL are for rendering in mobile apps. IN_STREAM_VIDEO refers to rendering in in-stream video ads developed with the VAST standard. Applicable to all creative types.

Acceptable values are: - "APP" - "APP_INTERSTITIAL" - "IN_STREAM_VIDEO" - "WEB" - "WEB_INTERSTITIAL"

ad_parameters

Ad parameters user for VPAID creative. This is a read-only field. Applicable to the following creative types: all VPAID.

ad_tag_keys

Keywords for a Rich Media creative. Keywords let you customize the creative settings of a Rich Media ad running on your site without having to contact the advertiser. You can use keywords to dynamically change the look or functionality of a creative. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#creative".

subaccount_id

Subaccount ID of this creative. This field, if left unset, will be auto-generated for both insert and update operations. Applicable to all creative types.

name

Name of the creative. This is a required field and must be less than 256 characters long. Applicable to all creative types.

convert_flash_to_html5

Whether Flash assets associated with the creative need to be automatically converted to HTML5. This flag is enabled by default and users can choose to disable it if they don't want the system to generate and use HTML5 asset for this creative. Applicable to the following creative types: ENHANCED_BANNER and FLASH_INPAGE.

video_duration

Creative video duration in seconds. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO, all RICH_MEDIA, and all VPAID.

last_modified_info

Creative last modification information. This is a read-only field. Applicable to all creative types.

backup_image_click_through_url

Click-through URL for backup image. Applicable to the following creative types: ENHANCED_BANNER, FLASH_INPAGE, and HTML5_BANNER.

studio_creative_id

Studio creative ID associated with rich media and VPAID creatives. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA, and all VPAID.

authoring_tool

Authoring tool for HTML5 banner creatives. This is a read-only field. Applicable to the following creative types: HTML5_BANNER.

third_party_urls

Third-party URLs for tracking in-stream video creative events. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

companion_creatives

List of companion creatives assigned to an in-Stream videocreative. Acceptable values include IDs of existing flash and image creatives. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

html_code

HTML code for the creative. This is a required field when applicable. This field is ignored if htmlCodeLocked is false. Applicable to the following creative types: all CUSTOM, FLASH_INPAGE, and HTML5_BANNER, and all RICH_MEDIA.

ssl_compliant

Whether the creative is SSL-compliant. This is a read-only field. Applicable to all creative types.

commercial_id

Industry standard ID assigned to creative for reach and frequency. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

skippable

Whether the user can choose to skip the creative. Applicable to the following creative types: INSTREAM_VIDEO.