Struct google_dfareporting2d1::CreativeAsset

pub struct CreativeAsset {
    pub mime_type: Option<String>,
    pub artwork_type: Option<String>,
    pub position_left_unit: Option<String>,
    pub collapsed_size: Option<Size>,
    pub flash_version: Option<i32>,
    pub pushdown: Option<bool>,
    pub hide_flash_objects: Option<bool>,
    pub detected_features: Option<Vec<String>>,
    pub duration: Option<i32>,
    pub position_top_unit: Option<String>,
    pub id: Option<String>,
    pub alignment: Option<String>,
    pub size: Option<Size>,
    pub horizontally_locked: Option<bool>,
    pub start_time_type: Option<String>,
    pub display_type: Option<String>,
    pub expanded_dimension: Option<Size>,
    pub vertically_locked: Option<bool>,
    pub role: Option<String>,
    pub zip_filesize: Option<String>,
    pub bit_rate: Option<i32>,
    pub window_mode: Option<String>,
    pub progressive_serving_url: Option<String>,
    pub duration_type: Option<String>,
    pub video_duration: Option<f32>,
    pub file_size: Option<String>,
    pub offset: Option<OffsetPosition>,
    pub active: Option<bool>,
    pub original_backup: Option<bool>,
    pub hide_selection_boxes: Option<bool>,
    pub streaming_serving_url: Option<String>,
    pub z_index: Option<i32>,
    pub child_asset_type: Option<String>,
    pub backup_image_exit: Option<CreativeCustomEvent>,
    pub action_script3: Option<bool>,
    pub pushdown_duration: Option<f32>,
    pub asset_identifier: Option<CreativeAssetId>,
    pub transparency: Option<bool>,
    pub ssl_compliant: Option<bool>,
    pub position: Option<OffsetPosition>,
    pub custom_start_time_value: Option<i32>,
    pub zip_filename: Option<String>,
}

Creative Asset.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert creative assets (none)

Fields

mime_type

Detected MIME type for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

artwork_type

Artwork type of rich media creative. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

position_left_unit

Offset left unit for an asset. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

collapsed_size

Size of an asset when collapsed. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA and all VPAID. Additionally, applicable to assets whose displayType is ASSET_DISPLAY_TYPE_EXPANDING or ASSET_DISPLAY_TYPE_PEEL_DOWN.

flash_version

Flash version of the asset. This is a read-only field. Applicable to the following creative types: FLASH_INPAGE, ENHANCED_BANNER, all RICH_MEDIA, and all VPAID.

pushdown

Whether the asset pushes down other content. Applicable to the following creative types: all RICH_MEDIA. Additionally, only applicable when the asset offsets are 0, the collapsedSize.width matches size.width, and the collapsedSize.height is less than size.height.

hide_flash_objects

Whether to hide Flash objects flag for an asset. Applicable to the following creative types: all RICH_MEDIA.

detected_features

List of feature dependencies for the creative asset that are detected by DCM. Feature dependencies are features that a browser must be able to support in order to render your HTML5 creative correctly. This is a read-only, auto-generated field. Applicable to the following creative types: ENHANCED_BANNER and HTML5_BANNER.

duration

Duration in seconds for which an asset will be displayed. Applicable to the following creative types: INSTREAM_VIDEO and VPAID_LINEAR.

position_top_unit

Offset top unit for an asset. This is a read-only field if the asset displayType is ASSET_DISPLAY_TYPE_OVERLAY. Applicable to the following creative types: all RICH_MEDIA.

id

Numeric ID of this creative asset. This is a required field and should not be modified. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT.

alignment

Possible alignments for an asset. This is a read-only field. Applicable to the following creative types: RICH_MEDIA_MULTI_FLOATING.

size

Size associated with this creative asset. This is a required field when applicable; however for IMAGE and FLASH_INPAGE creatives, if left blank, this field will be automatically set using the actual size of the associated image asset. Applicable to the following creative types: ENHANCED_BANNER, ENHANCED_IMAGE, FLASH_INPAGE, HTML5_BANNER, IMAGE, and all RICH_MEDIA.

horizontally_locked

Whether the asset is horizontally locked. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

start_time_type

Initial wait time type before making the asset visible. Applicable to the following creative types: all RICH_MEDIA.

display_type

Type of rich media asset. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

expanded_dimension

Detected expanded dimension for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

vertically_locked

Whether the asset is vertically locked. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.

role

Role of the asset in relation to creative. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT. This is a required field. PRIMARY applies to ENHANCED_BANNER, FLASH_INPAGE, HTML5_BANNER, IMAGE, IMAGE_GALLERY, all RICH_MEDIA (which may contain multiple primary assets), and all VPAID creatives. BACKUP_IMAGE applies to ENHANCED_BANNER, FLASH_INPAGE, HTML5_BANNER, all RICH_MEDIA, and all VPAID creatives. ADDITIONAL_IMAGE and ADDITIONAL_FLASH apply to FLASH_INPAGE creatives. OTHER refers to assets from sources other than DCM, such as Studio uploaded assets, applicable to all RICH_MEDIA and all VPAID creatives. PARENT_VIDEO refers to videos uploaded by the user in DCM and is applicable to INSTREAM_VIDEO and VPAID_LINEAR creatives. TRANSCODED_VIDEO refers to videos transcoded by DCM from PARENT_VIDEO assets and is applicable to INSTREAM_VIDEO and VPAID_LINEAR creatives. ALTERNATE_VIDEO refers to the DCM representation of child asset videos from Studio, and is applicable to VPAID_LINEAR creatives. These cannot be added or removed within DCM. For VPAID_LINEAR creatives, PARENT_VIDEO, TRANSCODED_VIDEO and ALTERNATE_VIDEO assets that are marked active serve as backup in case the VPAID creative cannot be served. Only PARENT_VIDEO assets can be added or removed for an INSTREAM_VIDEO or VPAID_LINEAR creative.

zip_filesize

Size of zip file. This is a read-only field. Applicable to the following creative types: HTML5_BANNER.

bit_rate

Detected bit-rate for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

window_mode

Window mode options for flash assets. Applicable to the following creative types: FLASH_INPAGE, RICH_MEDIA_EXPANDING, RICH_MEDIA_IM_EXPAND, RICH_MEDIA_INPAGE, and RICH_MEDIA_INPAGE_FLOATING.

progressive_serving_url

Progressive URL for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

duration_type

Duration type for which an asset will be displayed. Applicable to the following creative types: all RICH_MEDIA.

video_duration

Detected video duration for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

file_size

File size associated with this creative asset. This is a read-only field. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT.

offset

Offset position for an asset in collapsed mode. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA and all VPAID. Additionally, only applicable to assets whose displayType is ASSET_DISPLAY_TYPE_EXPANDING or ASSET_DISPLAY_TYPE_PEEL_DOWN.

active

Whether the video asset is active. This is a read-only field for VPAID_NON_LINEAR assets. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

original_backup

Whether the backup asset is original or changed by the user in DCM. Applicable to the following creative types: all RICH_MEDIA.

hide_selection_boxes

Whether to hide selection boxes flag for an asset. Applicable to the following creative types: all RICH_MEDIA.

streaming_serving_url

Streaming URL for video asset. This is a read-only field. Applicable to the following creative types: INSTREAM_VIDEO and all VPAID.

z_index

zIndex value of an asset. This is a read-only field. Applicable to the following creative types: all RICH_MEDIA.Additionally, only applicable to assets whose displayType is NOT one of the following types: ASSET_DISPLAY_TYPE_INPAGE or ASSET_DISPLAY_TYPE_OVERLAY.

child_asset_type

Rich media child asset type. This is a read-only field. Applicable to the following creative types: all VPAID.

backup_image_exit

Exit event configured for the backup image. Applicable to the following creative types: all RICH_MEDIA.

action_script3

Whether ActionScript3 is enabled for the flash asset. This is a read-only field. Applicable to the following creative types: FLASH_INPAGE and ENHANCED_BANNER.

pushdown_duration

Pushdown duration in seconds for an asset. Must be between 0 and 9.99. Applicable to the following creative types: all RICH_MEDIA.Additionally, only applicable when the asset pushdown field is true, the offsets are 0, the collapsedSize.width matches size.width, and the collapsedSize.height is less than size.height.

asset_identifier

Identifier of this asset. This is the same identifier returned during creative asset insert operation. This is a required field. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT.

transparency

Whether the asset is transparent. Applicable to the following creative types: all RICH_MEDIA. Additionally, only applicable to HTML5 assets.

ssl_compliant

Whether the asset is SSL-compliant. This is a read-only field. Applicable to all but the following creative types: all REDIRECT and TRACKING_TEXT.

position

Offset position for an asset. Applicable to the following creative types: all RICH_MEDIA.

custom_start_time_value

Custom start time in seconds for making the asset visible. Applicable to the following creative types: all RICH_MEDIA.

zip_filename

File name of zip file. This is a read-only field. Applicable to the following creative types: HTML5_BANNER.