Struct google_dfareporting2d1::CreativeCustomEvent

pub struct CreativeCustomEvent {
    pub target_type: Option<String>,
    pub artwork_type: Option<String>,
    pub video_reporting_id: Option<String>,
    pub popup_window_properties: Option<PopupWindowProperties>,
    pub artwork_label: Option<String>,
    pub active: Option<bool>,
    pub advertiser_custom_event_type: Option<String>,
    pub exit_url: Option<String>,
    pub id: Option<String>,
    pub advertiser_custom_event_name: Option<String>,
}

Creative Custom Event.

This type is not used in any activity, and only used as part of another schema.

Fields

target_type

Target type used by the event.

artwork_type

Artwork type used by the creative.This is a read-only field.

video_reporting_id

Reporting ID, used to differentiate multiple videos in a single creative.

popup_window_properties

Properties for rich media popup windows. This field is used only for exit events.

artwork_label

Artwork label column, used to link events in DCM back to events in Studio. This is a required field and should not be modified after insertion.

active

Whether the event is active.

advertiser_custom_event_type

Type of the event. This is a read-only field.

exit_url

Exit URL of the event. This field is used only for exit events.

id

ID of this event. This is a required field and should not be modified after insertion.

advertiser_custom_event_name

User-entered name for the event.