Struct google_dfareporting2d1::EventTagsListResponse

pub struct EventTagsListResponse {
    pub event_tags: Option<Vec<EventTag>>,
    pub kind: Option<String>,
}

Event Tag List Response

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list event tags (response)

Fields

event_tags

Event tag collection.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#eventTagsListResponse".