Struct google_dfareporting2d1::FloodlightActivityGroup

pub struct FloodlightActivityGroup {
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub floodlight_configuration_id: Option<String>,
    pub advertiser_id: Option<String>,
    pub floodlight_configuration_id_dimension_value: Option<DimensionValue>,
    pub id_dimension_value: Option<DimensionValue>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub tag_string: Option<String>,
    pub type_: Option<String>,
    pub id: Option<String>,
    pub account_id: Option<String>,
}

Contains properties of a Floodlight activity group.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get floodlight activity groups (response)

patch floodlight activity groups (request|response)

update floodlight activity groups (request|response)

list floodlight activity groups (none)

delete floodlight activity groups (none)

insert floodlight activity groups (request|response)

Fields

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#floodlightActivityGroup".

subaccount_id

Subaccount ID of this floodlight activity group. This is a read-only field that can be left blank.

name

Name of this floodlight activity group. This is a required field. Must be less than 65 characters long and cannot contain quotes.

floodlight_configuration_id

Floodlight configuration ID of this floodlight activity group. This is a required field.

advertiser_id

Advertiser ID of this floodlight activity group. If this field is left blank, the value will be copied over either from the floodlight configuration's advertiser or from the existing activity group's advertiser.

floodlight_configuration_id_dimension_value

Dimension value for the ID of the floodlight configuration. This is a read-only, auto-generated field.

id_dimension_value

Dimension value for the ID of this floodlight activity group. This is a read-only, auto-generated field.

advertiser_id_dimension_value

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

tag_string

Value of the type= parameter in the floodlight tag, which the ad servers use to identify the activity group that the activity belongs to. This is optional: if empty, a new tag string will be generated for you. This string must be 1 to 8 characters long, with valid characters being [a-z][A-Z][0-9][-][ _ ]. This tag string must also be unique among activity groups of the same floodlight configuration. This field is read-only after insertion.

type_

Type of the floodlight activity group. This is a required field that is read-only after insertion.

id

ID of this floodlight activity group. This is a read-only, auto-generated field.

account_id

Account ID of this floodlight activity group. This is a read-only field that can be left blank.