Struct google_dfareporting2d1::FloodlightActivityPublisherDynamicTag

pub struct FloodlightActivityPublisherDynamicTag {
    pub view_through: Option<bool>,
    pub site_id: Option<String>,
    pub click_through: Option<bool>,
    pub directory_site_id: Option<String>,
    pub site_id_dimension_value: Option<DimensionValue>,
    pub dynamic_tag: Option<FloodlightActivityDynamicTag>,
}

Publisher Dynamic Tag

This type is not used in any activity, and only used as part of another schema.

Fields
view_through: Option<bool>

Whether this tag is applicable only for view-throughs.

site_id: Option<String>

Site ID of this dynamic tag.

click_through: Option<bool>

Whether this tag is applicable only for click-throughs.

directory_site_id: Option<String>

Directory site ID of this dynamic tag. This is a write-only field that can be used as an alternative to the siteId field. When this resource is retrieved, only the siteId field will be populated.

site_id_dimension_value: Option<DimensionValue>

Dimension value for the ID of the site. This is a read-only, auto-generated field.

dynamic_tag: Option<FloodlightActivityDynamicTag>

Dynamic floodlight tag.