Struct google_dfareporting2d1::InventoryItem

pub struct InventoryItem {
    pub order_id: Option<String>,
    pub advertiser_id: Option<String>,
    pub site_id: Option<String>,
    pub ad_slots: Option<Vec<AdSlot>>,
    pub rfp_id: Option<String>,
    pub content_category_id: Option<String>,
    pub placement_strategy_id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub negotiation_channel_id: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub estimated_click_through_rate: Option<String>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub estimated_conversion_rate: Option<String>,
    pub in_plan: Option<bool>,
    pub id: Option<String>,
    pub project_id: Option<String>,
    pub pricing: Option<Pricing>,
}

Represents a buy from the DoubleClick Planning inventory store.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get inventory items (response)

list inventory items (none)

Fields

order_id

Order ID of this inventory item.

advertiser_id

Advertiser ID of this inventory item.

site_id

ID of the site this inventory item is associated with.

ad_slots

Ad slots of this inventory item. If this inventory item represents a standalone placement, there will be exactly one ad slot. If this inventory item represents a placement group, there will be more than one ad slot, each representing one child placement in that placement group.

rfp_id

RFP ID of this inventory item.

content_category_id

Content category ID of this inventory item.

placement_strategy_id

Placement strategy ID of this inventory item.

account_id

Account ID of this inventory item.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#inventoryItem".

negotiation_channel_id

Negotiation channel ID of this inventory item.

subaccount_id

Subaccount ID of this inventory item.

name

Name of this inventory item. For standalone inventory items, this is the same name as that of its only ad slot. For group inventory items, this can differ from the name of any of its ad slots.

estimated_click_through_rate

Estimated click-through rate of this inventory item.

last_modified_info

Information about the most recent modification of this inventory item.

estimated_conversion_rate

Estimated conversion rate of this inventory item.

in_plan

Whether this inventory item is in plan.

id

ID of this inventory item.

project_id

Project ID of this inventory item.

pricing

Pricing of this inventory item.