Struct google_dfareporting2d1::ListPopulationTerm

pub struct ListPopulationTerm {
    pub remarketing_list_id: Option<String>,
    pub contains: Option<bool>,
    pub negation: Option<bool>,
    pub value: Option<String>,
    pub variable_friendly_name: Option<String>,
    pub operator: Option<String>,
    pub type_: Option<String>,
    pub variable_name: Option<String>,
}

Remarketing List Population Rule Term.

This type is not used in any activity, and only used as part of another schema.

Fields

remarketing_list_id

ID of the list in question. This field is only relevant when type is set to LIST_MEMBERSHIP_TERM.

contains

Will be true if the term should check if the user is in the list and false if the term should check if the user is not in the list. This field is only relevant when type is set to LIST_MEMBERSHIP_TERM. False by default.

negation

Whether to negate the comparison result of this term during rule evaluation. This field is only relevant when type is left unset or set to CUSTOM_VARIABLE_TERM or REFERRER_TERM.

value

Literal to compare the variable to. This field is only relevant when type is left unset or set to CUSTOM_VARIABLE_TERM or REFERRER_TERM.

variable_friendly_name

Friendly name of this term's variable. This is a read-only, auto-generated field. This field is only relevant when type is left unset or set to CUSTOM_VARIABLE_TERM.

operator

Comparison operator of this term. This field is only relevant when type is left unset or set to CUSTOM_VARIABLE_TERM or REFERRER_TERM.

type_

List population term type determines the applicable fields in this object. If left unset or set to CUSTOM_VARIABLE_TERM, then variableName, variableFriendlyName, operator, value, and negation are applicable. If set to LIST_MEMBERSHIP_TERM then remarketingListId and contains are applicable. If set to REFERRER_TERM then operator, value, and negation are applicable.

variable_name

Name of the variable (U1, U2, etc.) being compared in this term. This field is only relevant when type is set to null, CUSTOM_VARIABLE_TERM or REFERRER_TERM.