Struct google_dfareporting2d1::OperatingSystemVersion

pub struct OperatingSystemVersion {
    pub major_version: Option<String>,
    pub kind: Option<String>,
    pub minor_version: Option<String>,
    pub name: Option<String>,
    pub operating_system: Option<OperatingSystem>,
    pub id: Option<String>,
}

Contains information about a particular version of an operating system that can be targeted by ads.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get operating system versions (response)

list operating system versions (none)

Fields
major_version: Option<String>

Major version (leftmost number) of this operating system version.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#operatingSystemVersion".

minor_version: Option<String>

Minor version (number after the first dot) of this operating system version.

name: Option<String>

Name of this operating system version.

operating_system: Option<OperatingSystem>

Operating system of this operating system version.

id: Option<String>

ID of this operating system version.