Struct google_dfareporting2d1::Order

pub struct Order {
    pub terms_and_conditions: Option<String>,
    pub buyer_organization_name: Option<String>,
    pub advertiser_id: Option<String>,
    pub site_id: Option<Vec<String>>,
    pub buyer_invoice_id: Option<String>,
    pub planning_term_id: Option<String>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub contacts: Option<Vec<OrderContact>>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub notes: Option<String>,
    pub seller_organization_name: Option<String>,
    pub comments: Option<String>,
    pub approver_user_profile_ids: Option<Vec<String>>,
    pub seller_order_id: Option<String>,
    pub project_id: Option<String>,
    pub site_names: Option<Vec<String>>,
}

Describes properties of a DoubleClick Planning order.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list orders (none)

get orders (response)

Fields

terms_and_conditions

Terms and conditions of this order.

buyer_organization_name

Name of the buyer organization.

advertiser_id

Advertiser ID of this order.

site_id

Site IDs this order is associated with.

buyer_invoice_id

Buyer invoice ID associated with this order.

planning_term_id

ID of the terms and conditions template used in this order.

id

ID of this order. This is a read-only, auto-generated field.

account_id

Account ID of this order.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#order".

subaccount_id

Subaccount ID of this order.

name

Name of this order.

contacts

Contacts for this order.

last_modified_info

Information about the most recent modification of this order.

notes

Notes of this order.

seller_organization_name

Name of the seller organization.

comments

Comments in this order.

approver_user_profile_ids

IDs for users that have to approve documents created for this order.

seller_order_id

Seller order ID associated with this order.

project_id

Project ID of this order.

site_names

Free-form site names this order is associated with.