Struct google_dfareporting2d1::PlacementGroup

pub struct PlacementGroup {
    pub comment: Option<String>,
    pub primary_placement_id: Option<String>,
    pub campaign_id: Option<String>,
    pub primary_placement_id_dimension_value: Option<DimensionValue>,
    pub directory_site_id_dimension_value: Option<DimensionValue>,
    pub advertiser_id: Option<String>,
    pub directory_site_id: Option<String>,
    pub create_info: Option<LastModifiedInfo>,
    pub site_id: Option<String>,
    pub pricing_schedule: Option<PricingSchedule>,
    pub site_id_dimension_value: Option<DimensionValue>,
    pub placement_group_type: Option<String>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub archived: Option<bool>,
    pub id: Option<String>,
    pub placement_strategy_id: Option<String>,
    pub account_id: Option<String>,
    pub programmatic_setting: Option<ProgrammaticSetting>,
    pub campaign_id_dimension_value: Option<DimensionValue>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub child_placement_ids: Option<Vec<String>>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub kind: Option<String>,
    pub content_category_id: Option<String>,
    pub external_id: Option<String>,
    pub id_dimension_value: Option<DimensionValue>,
}

Contains properties of a package or roadblock.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update placement groups (request|response)

patch placement groups (request|response)

get placement groups (response)

insert placement groups (request|response)

list placement groups (none)

Fields

comment

Comments for this placement group.

primary_placement_id

ID of the primary placement, used to calculate the media cost of a roadblock (placement group). Modifying this field will automatically modify the primary field on all affected roadblock child placements.

campaign_id

Campaign ID of this placement group. This field is required on insertion.

primary_placement_id_dimension_value

Dimension value for the ID of the primary placement. This is a read-only, auto-generated field.

directory_site_id_dimension_value

Dimension value for the ID of the directory site. This is a read-only, auto-generated field.

advertiser_id

Advertiser ID of this placement group. This is a required field on insertion.

directory_site_id

Directory site ID associated with this placement group. On insert, you must set either this field or the site_id field to specify the site associated with this placement group. This is a required field that is read-only after insertion.

create_info

Information about the creation of this placement group. This is a read-only field.

site_id

Site ID associated with this placement group. On insert, you must set either this field or the directorySiteId field to specify the site associated with this placement group. This is a required field that is read-only after insertion.

pricing_schedule

Pricing schedule of this placement group. This field is required on insertion.

site_id_dimension_value

Dimension value for the ID of the site. This is a read-only, auto-generated field.

placement_group_type

Type of this placement group. A package is a simple group of placements that acts as a single pricing point for a group of tags. A roadblock is a group of placements that not only acts as a single pricing point, but also assumes that all the tags in it will be served at the same time. A roadblock requires one of its assigned placements to be marked as primary for reporting. This field is required on insertion.

advertiser_id_dimension_value

Dimension value for the ID of the advertiser. This is a read-only, auto-generated field.

archived

Whether this placement group is archived.

id

ID of this placement group. This is a read-only, auto-generated field.

placement_strategy_id

ID of the placement strategy assigned to this placement group.

account_id

Account ID of this placement group. This is a read-only field that can be left blank.

programmatic_setting

Settings for a programmatic placement.

campaign_id_dimension_value

Dimension value for the ID of the campaign. This is a read-only, auto-generated field.

subaccount_id

Subaccount ID of this placement group. This is a read-only field that can be left blank.

name

Name of this placement group. This is a required field and must be less than 256 characters long.

child_placement_ids

IDs of placements which are assigned to this placement group. This is a read-only, auto-generated field.

last_modified_info

Information about the most recent modification of this placement group. This is a read-only field.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#placementGroup".

content_category_id

ID of the content category assigned to this placement group.

external_id

External ID for this placement.

id_dimension_value

Dimension value for the ID of this placement group. This is a read-only, auto-generated field.