Struct google_dfareporting2d1::ReportMethods

pub struct ReportMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on report resources. It is not used directly, but through the Dfareporting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_dfareporting2d1 as dfareporting2d1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2d1::Dfareporting;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `compatible_fields_query(...)`, `delete(...)`, `files_get(...)`, `files_list(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)`, `run(...)` and `update(...)`
// to build up your call.
let rb = hub.reports();

Methods

impl<'a, C, A> ReportMethods<'a, C, A>

fn run(&self, profile_id: &str, report_id: &str) -> ReportRunCall<'a, C, A>

Create a builder to help you perform the following task:

Runs a report.

Arguments

profileId - The DFA profile ID.

reportId - The ID of the report.

fn patch(&self, request: Report, profile_id: &str, report_id: &str) -> ReportPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a report. This method supports patch semantics.

Arguments

request - No description provided.

profileId - The DFA user profile ID.

reportId - The ID of the report.

fn files_list(&self, profile_id: &str, report_id: &str) -> ReportFileListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists files for a report.

Arguments

profileId - The DFA profile ID.

reportId - The ID of the parent report.

fn insert(&self, request: Report, profile_id: &str) -> ReportInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a report.

Arguments

request - No description provided.

profileId - The DFA user profile ID.

fn update(&self, request: Report, profile_id: &str, report_id: &str) -> ReportUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a report.

Arguments

request - No description provided.

profileId - The DFA user profile ID.

reportId - The ID of the report.

fn compatible_fields_query(&self, request: Report, profile_id: &str) -> ReportCompatibleFieldQueryCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the fields that are compatible to be selected in the respective sections of a report criteria, given the fields already selected in the input report and user permissions.

Arguments

request - No description provided.

profileId - The DFA user profile ID.

fn get(&self, profile_id: &str, report_id: &str) -> ReportGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a report by its ID.

Arguments

profileId - The DFA user profile ID.

reportId - The ID of the report.

fn files_get(&self, profile_id: &str, report_id: &str, file_id: &str) -> ReportFileGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a report file.

Arguments

profileId - The DFA profile ID.

reportId - The ID of the report.

fileId - The ID of the report file.

fn delete(&self, profile_id: &str, report_id: &str) -> ReportDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a report by its ID.

Arguments

profileId - The DFA user profile ID.

reportId - The ID of the report.

fn list(&self, profile_id: &str) -> ReportListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves list of reports.

Arguments

profileId - The DFA user profile ID.