Struct google_dfareporting2d1::ReportPathToConversionCriteria

pub struct ReportPathToConversionCriteria {
    pub conversion_dimensions: Option<Vec<SortedDimension>>,
    pub metric_names: Option<Vec<String>>,
    pub date_range: Option<DateRange>,
    pub custom_floodlight_variables: Option<Vec<SortedDimension>>,
    pub activity_filters: Option<Vec<DimensionValue>>,
    pub custom_rich_media_events: Option<Vec<DimensionValue>>,
    pub per_interaction_dimensions: Option<Vec<SortedDimension>>,
    pub report_properties: Option<ReportPathToConversionCriteriaReportProperties>,
    pub floodlight_config_id: Option<DimensionValue>,
}

The report criteria for a report of type "PATH_TO_CONVERSION".

This type is not used in any activity, and only used as part of another schema.

Fields

conversion_dimensions

The list of conversion dimensions the report should include.

metric_names

The list of names of metrics the report should include.

date_range

The date range this report should be run for.

custom_floodlight_variables

The list of custom floodlight variables the report should include.

activity_filters

The list of 'dfa:activity' values to filter on.

custom_rich_media_events

The list of custom rich media events to include.

per_interaction_dimensions

The list of per interaction dimensions the report should include.

report_properties

The properties of the report.

floodlight_config_id

The floodlight ID for which to show data in this report. All advertisers associated with that ID will automatically be added. The dimension of the value needs to be 'dfa:floodlightConfigId'.