Struct google_dfareporting2d1::AccountUserProfile

pub struct AccountUserProfile {
    pub locale: Option<String>,
    pub site_filter: Option<ObjectFilter>,
    pub campaign_filter: Option<ObjectFilter>,
    pub user_role_id: Option<String>,
    pub user_access_type: Option<String>,
    pub active: Option<bool>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub user_role_filter: Option<ObjectFilter>,
    pub trafficker_type: Option<String>,
    pub comments: Option<String>,
    pub name: Option<String>,
    pub advertiser_filter: Option<ObjectFilter>,
    pub email: Option<String>,
}

AccountUserProfiles contains properties of a DCM user profile. This resource is specifically for managing user profiles, whereas UserProfiles is for accessing the API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch account user profiles (request|response)

list account user profiles (none)

insert account user profiles (request|response)

get account user profiles (response)

update account user profiles (request|response)

Fields

locale

Locale of the user profile. This is a required field. Acceptable values are:
- "cs" (Czech) - "de" (German) - "en" (English) - "en-GB" (English United Kingdom) - "es" (Spanish) - "fr" (French) - "it" (Italian) - "ja" (Japanese) - "ko" (Korean) - "pl" (Polish) - "pt-BR" (Portuguese Brazil) - "ru" (Russian) - "sv" (Swedish) - "tr" (Turkish) - "zh-CN" (Chinese Simplified) - "zh-TW" (Chinese Traditional)

site_filter

Filter that describes which sites are visible to the user profile.

campaign_filter

Filter that describes which campaigns are visible to the user profile.

user_role_id

User role ID of the user profile. This is a required field.

user_access_type

User type of the user profile. This is a read-only field that can be left blank.

active

Whether this user profile is active. This defaults to false, and must be set true on insert for the user profile to be usable.

id

ID of the user profile. This is a read-only, auto-generated field.

account_id

Account ID of the user profile. This is a read-only field that can be left blank.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#accountUserProfile".

subaccount_id

Subaccount ID of the user profile. This is a read-only field that can be left blank.

user_role_filter

Filter that describes which user roles are visible to the user profile.

trafficker_type

Trafficker type of this user profile.

comments

Comments for this user profile.

name

Name of the user profile. This is a required field. Must be less than 64 characters long, must be globally unique, and cannot contain whitespace or any of the following characters: "&;"#%,".

advertiser_filter

Filter that describes which advertisers are visible to the user profile.

email

Email of the user profile. The email addresss must be linked to a Google Account. This field is required on insertion and is read-only after insertion.