Struct google_dfareporting2d1::CreativeAssetMetadata

pub struct CreativeAssetMetadata {
    pub warned_validation_rules: Option<Vec<String>>,
    pub detected_features: Option<Vec<String>>,
    pub click_tags: Option<Vec<ClickTag>>,
    pub asset_identifier: Option<CreativeAssetId>,
    pub kind: Option<String>,
}

CreativeAssets contains properties of a creative asset file which will be uploaded or has already been uploaded. Refer to the creative sample code for how to upload assets and insert a creative.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert creative assets (request|response)

Fields
warned_validation_rules: Option<Vec<String>>

Rules validated during code generation that generated a warning. This is a read-only, auto-generated field.

Possible values are: - "CLICK_TAG_NON_TOP_LEVEL" - "CLICK_TAG_MISSING" - "CLICK_TAG_MORE_THAN_ONE" - "CLICK_TAG_INVALID" - "ORPHANED_ASSET" - "PRIMARY_HTML_MISSING" - "EXTERNAL_FILE_REFERENCED" - "MRAID_REFERENCED" - "ADMOB_REFERENCED" - "FILE_TYPE_INVALID" - "ZIP_INVALID" - "LINKED_FILE_NOT_FOUND" - "MAX_FLASH_VERSION_11" - "NOT_SSL_COMPLIANT" - "FILE_DETAIL_EMPTY" - "ASSET_INVALID" - "GWD_PROPERTIES_INVALID" - "ENABLER_UNSUPPORTED_METHOD_DCM" - "ASSET_FORMAT_UNSUPPORTED_DCM" - "COMPONENT_UNSUPPORTED_DCM" - "HTML5_FEATURE_UNSUPPORTED' "

detected_features: Option<Vec<String>>

List of feature dependencies for the creative asset that are detected by DCM. Feature dependencies are features that a browser must be able to support in order to render your HTML5 creative correctly. This is a read-only, auto-generated field.

click_tags: Option<Vec<ClickTag>>

List of detected click tags for assets. This is a read-only auto-generated field.

asset_identifier: Option<CreativeAssetId>

ID of the creative asset. This is a required field.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#creativeAssetMetadata".