Struct google_dfareporting2d1::OrderDocument

pub struct OrderDocument {
    pub order_id: Option<String>,
    pub advertiser_id: Option<String>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub effective_date: Option<String>,
    pub title: Option<String>,
    pub project_id: Option<String>,
    pub amended_order_document_id: Option<String>,
    pub signed: Option<bool>,
    pub created_info: Option<LastModifiedInfo>,
    pub approved_by_user_profile_ids: Option<Vec<String>>,
    pub cancelled: Option<bool>,
    pub type_: Option<String>,
    pub id: Option<String>,
    pub account_id: Option<String>,
}

Contains properties of a DoubleClick Planning order document.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get order documents (response)

list order documents (none)

Fields
order_id: Option<String>

ID of the order from which this order document is created.

advertiser_id: Option<String>

Advertiser ID of this order document.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#orderDocument".

subaccount_id: Option<String>

Subaccount ID of this order document.

effective_date: Option<String>

Effective date of this order document.

title: Option<String>

Title of this order document.

project_id: Option<String>

Project ID of this order document.

amended_order_document_id: Option<String>

The amended order document ID of this order document. An order document can be created by optionally amending another order document so that the change history can be preserved.

signed: Option<bool>

Whether this order document has been signed.

created_info: Option<LastModifiedInfo>

Information about the creation of this order document.

approved_by_user_profile_ids: Option<Vec<String>>

IDs of users who have approved this order document.

cancelled: Option<bool>

Whether this order document is cancelled.

type_: Option<String>

Type of this order document

id: Option<String>

ID of this order document.

account_id: Option<String>

Account ID of this order document.